+++ Dockerfile
@@ -2,20 +2,24 @@
 # nodejs 20 hangs on build with armv6/armv7
 FROM docker.io/library/node:18-alpine AS build_node_modules
 
-# Update npm to latest
-RUN npm install -g npm@latest
+# Install pnpm
+RUN corepack enable pnpm
 
 # Copy Web UI
 COPY src /app
 WORKDIR /app
-RUN npm ci --omit=dev &&\
-    mv node_modules /node_modules
+RUN pnpm install --prod
+
+# Build UI
+RUN pnpm build
+
+RUN mv node_modules /node_modules
 
 # Copy build result to a new image.
 # This saves a lot of disk space.
 FROM docker.io/library/node:20-alpine
 HEALTHCHECK CMD /usr/bin/timeout 5s /bin/sh -c "/usr/bin/wg show | /bin/grep -q interface || exit 1" --interval=1m --timeout=5s --retries=3
-COPY --from=build_node_modules /app /app
+COPY --from=build_node_modules /app/.output /app
 
 # Move node_modules one directory up, so during development
 # we don't have to mount it in a volume.
@@ -46,4 +50,4 @@ ENV DEBUG=Server,WireGuard
 
 # Run Web UI
 WORKDIR /app
-CMD ["/usr/bin/dumb-init", "node", "server.js"]
+CMD ["/usr/bin/dumb-init", "node", "server/index.mjs"]
