--- src/lib/ServerError.js
@@ -1,10 +1,0 @@
-'use strict';
-
-module.exports = class ServerError extends Error {
-
-  constructor(message, statusCode = 500) {
-    super(message);
-    this.statusCode = statusCode;
-  }
-
-};
