+++ src/middleware/session.ts
@@ -1,0 +1,5 @@
+export default defineNuxtRouteMiddleware(async (to, from) => {
+    if (REQUIRES_PASSWORD || !to.path.startsWith('/api/')) {
+        return abortNavigation();
+    }
+  })
