+++ src/nuxt.config.ts
@@ -2,5 +2,5 @@
 export default defineNuxtConfig({
   compatibilityDate: '2024-04-03',
   devtools: { enabled: true },
-  modules: ["@nuxtjs/i18n"]
+  modules: ["@nuxtjs/i18n", "@nuxtjs/tailwindcss"]
 })
