+++ src/package.json
@@ -5,33 +5,28 @@
   "name": "wg-easy",
   "version": "1.0.1",
   "description": "The easiest way to run WireGuard VPN + Web-based Admin UI.",
-  "main": "server.js",
+  "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "DEBUG=Server,WireGuard npx nodemon server.js",
-    "serve-with-password": "PASSWORD=wg npm run serve",
-    "lint": "eslint .",
-    "buildcss": "npx tailwindcss -i ./www/src/css/app.css -o ./www/css/app.css"
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare"
   },
-  "author": "Emile Nijssen",
-  "license": "CC BY-NC-SA 4.0",
   "dependencies": {
-    "bcryptjs": "^2.4.3",
+    "@nuxtjs/i18n": "^8.3.3",
+    "@nuxtjs/tailwindcss": "^6.12.1",
+    "bcrypt": "^5.1.1",
     "debug": "^4.3.6",
-    "express-session": "^1.18.0",
-    "h3": "^1.12.0",
-    "qrcode": "^1.5.3"
+    "nuxt": "^3.12.4",
+    "qrcode": "^1.5.3",
+    "vue": "latest"
   },
   "devDependencies": {
-    "eslint-config-athom": "^3.1.3",
-    "nodemon": "^3.1.4",
-    "tailwindcss": "^3.4.7"
+    "@types/bcrypt": "^5.0.2",
+    "@types/debug": "^4.1.12",
+    "@types/qrcode": "^1.5.5"
   },
-  "nodemonConfig": {
-    "ignore": [
-      "www/*"
-    ]
-  },
-  "engines": {
-    "node": ">=18"
-  }
+  "packageManager": "pnpm@8.15.4+sha1.c85a4305534f76d461407b59277b954bac97b5c4"
 }
