+++ src/server/api/lang.ts
@@ -1,0 +1,6 @@
+export default defineEventHandler((event) => {
+    assertMethod(event, "GET");
+    const {LANG} = useRuntimeConfig();
+    setHeader(event, 'Content-Type', 'application/json');
+    return `"${LANG}"`;
+})
