+++ src/server/api/release.ts
@@ -1,0 +1,6 @@
+export default defineEventHandler((event) => {
+    assertMethod(event, "GET");
+    const {RELEASE} = useRuntimeConfig();
+    setHeader(event, 'Content-Type', 'application/json');
+    return RELEASE;
+})
