+++ src/server/api/ui-traffic-stats.ts
@@ -1,0 +1,6 @@
+export default defineEventHandler((event) => {
+    assertMethod(event, "GET");
+    const {UI_TRAFFIC_STATS} = useRuntimeConfig();
+    setHeader(event, 'Content-Type', 'application/json');
+    return `"${UI_TRAFFIC_STATS}"`;
+})
