+++ src/server/api/wireguard/client.ts
@@ -1,0 +1,5 @@
+import WireGuard from "~/utils/WireGuard";
+
+export default defineEventHandler(() => {
+    return WireGuard.getClients();
+})
