--- src/services/Server.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const Server = require('../lib/Server');
-
-module.exports = new Server();
