--- src/services/WireGuard.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const WireGuard = require('../lib/WireGuard');
-
-module.exports = new WireGuard();
