+++ src/utils/cmd.ts
@@ -1,0 +1,26 @@
+import childProcess from 'child_process';
+
+export function exec(cmd, {
+    log = true,
+  } = {}) {
+    if (typeof log === 'string') {
+      // eslint-disable-next-line no-console
+      console.log(`$ ${log}`);
+    } else if (log === true) {
+      // eslint-disable-next-line no-console
+      console.log(`$ ${cmd}`);
+    }
+
+    if (process.platform !== 'linux') {
+      return Promise.resolve("");
+    }
+
+    return new Promise((resolve, reject) => {
+      childProcess.exec(cmd, {
+        shell: 'bash',
+      }, (err, stdout) => {
+        if (err) return reject(err);
+        return resolve(String(stdout).trim());
+      });
+    });
+  }
