+++ src/utils/ip.ts
@@ -1,0 +1,12 @@
+export function isValidIPv4(str) {
+    const blocks = str.split('.');
+    if (blocks.length !== 4) return false;
+
+    for (let value of blocks) {
+      value = parseInt(value, 10);
+      if (Number.isNaN(value)) return false;
+      if (value < 0 || value > 255) return false;
+    }
+
+    return true;
+  }
