--- src/www/manifest.json
@@ -1,11 +1,0 @@
-{
-  "name": "WireGuard",
-  "display": "standalone",
-  "background_color": "#fff",
-  "icons": [
-    {
-      "src": "img/favicon.png",
-      "type": "image/png"
-    }
-  ]
-}
